fix unit tests and id types

+++ src/main/java/tn/esprit/spring/services/IEmployeService.java
@@ -33,7 +33,7 @@ public interface IEmployeService {
     Date dateDebut, Date dateFin);
 
     int addOrUpdateEmploye(Employe employe);
-    public long ajouterEmploye(Employe employe);
+    public int ajouterEmploye(Employe employe);
